Rename Redux.store.ts to Redux.slice.ts
